requirements.txt
pydantic>=2.0,<3.0
# Models should be compatible with Pydantic V2.x.
# Ensure V1 specific features (like old constr) are updated.
requests>=2.20.0
playwright>=1.20.0
google-adk
google-generativeai
